Update raster_intro.Rmd
There is no "format change" involved. "raster" reads (via ncdf4) and interprets netCDF files (like it would read a tif or any other format, except that these are accessed via gdal) . Of course interpretation could go wrong (more than anything because ncdf is very flexible, and creators of the files users are not forced to follow standards). Also, coordinates in "raster" refer to the center of grid cells --- as the examples below show. The "extent" of a raster, however, refers to the four outer cell corners.